requirements/base.txt
alabaster==0.7.11
asgiref==2.3.2
astroid==1.6.0
async-timeout==3.0.1
attrs==18.2.0
authy==2.1.5
autobahn==19.2.1
Automat==0.7.0
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cloudinary==1.15.0
channels>=2.1.1  # https://github.com/django/channels
channels-redis>=2.2.1  # https://github.com/django/channels_redis
colorama==0.3.9
constantly==15.1.0
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==2.2.0
django-filter==2.0.0
djangorestframework==3.8.2
docutils==0.14
future==0.16.0
geocoder==1.38.1
googlemaps==3.0.2
gunicorn==19.8.1
hyperlink==18.0.0
idna==2.7
imagesize==1.0.0
IMAPClient==2.0.0
incremental==17.5.0
isort==4.2.15
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
oauthlib==2.0.6
packaging==17.1
pbr==5.1.1
Pillow==5.1.0
psycopg2==2.7.5
pycryptodome==3.7.3
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.1
pylint==1.8.1
pyparsing==2.2.0
python-decouple==3.1
python-http-client==3.1.0
python-memcached==1.59
python-twitter==3.4
pytz==2018.4
ratelim==0.1.6
requests==2.19.1
requests-oauthlib==0.8.0
sendgrid==3.6.5
sendgrid-django==4.2.0
simplejson==3.15.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
txaio==18.8.1
urllib3==1.23
whitenoise==3.3.1
wrapt==1.10.11
zope.interface==4.6.0
